app/requirements.txt
wheel~=0.37.1
Flask~=2.0.2
Flask-Cors~=3.0.10
gunicorn==20.1.0
requests==2.26.0
Werkzeug==2.0.3
psycopg2
psycopg2-binary
SQLAlchemy~=1.4.31
Flask-SQLAlchemy~=2.5.1
Jinja2==3.0.3
python-dotenv==1.1.1
pandas~=2.3.1
trueskill==0.4.5
waitress==2.1.2
alembic==1.12.0